+++ src/main/java/com/ting/ting/repository/GroupDateRequestRepository.java
@@ -2,6 +2,9 @@ package com.ting.ting.repository;
 
 import com.ting.ting.domain.Group;
 import com.ting.ting.domain.GroupDateRequest;
+import com.ting.ting.domain.custom.GroupIdWithLikeCount;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -17,6 +20,8 @@ public interface GroupDateRequestRepository extends JpaRepository<GroupDateReque
     void deleteByFromGroup_IdAndToGroup_Id(Long fromGroupId, Long toGroupId);
 
     List<GroupDateRequest> findAllByFromGroup(Group fromGroup);
+
+    Page<GroupDateRequest> findAllByToGroup(Group fromGroup, Pageable pageable);
 
     @Query(value = "select distinct entity.fromGroup from GroupDateRequest entity join entity.fromGroup where entity.toGroup = :toGroup")
     List<Group> findFromGroupByToGroup(@Param("toGroup") Group toGroup);
